Added MFA

+++ src/main/java/com/francis/bookshop/controller/BookController.java
@@ -1,10 +1,12 @@
 package com.francis.bookshop.controller;
 
 import com.francis.bookshop.entity.Book;
+import com.francis.bookshop.service.AuthService;
 import com.francis.bookshop.service.BookService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -21,21 +23,25 @@ import java.util.List;
 @RequiredArgsConstructor
 public class BookController {
    private final BookService bookService;
+   private final AuthService authService;
     @PostMapping
     @PreAuthorize("hasRole('ADMIN')")
-    public ResponseEntity<Book> addBook(@RequestBody Book book) {
+    public ResponseEntity<Book> addBook(@RequestBody Book book, Authentication auth) {
+        authService.ensureAdmin(auth.getName());
         return ResponseEntity.ok(bookService.addBook(book));
     }
 
     @PutMapping("/{id}")
     @PreAuthorize("hasRole('ADMIN')")
-    public ResponseEntity<Book> updateBook(@PathVariable Long id, @RequestBody Book book) {
+    public ResponseEntity<Book> updateBook(@PathVariable Long id, @RequestBody Book book,  Authentication auth) {
+        authService.ensureAdmin(auth.getName());
         return ResponseEntity.ok(bookService.updateBook(id, book));
     }
 
     @DeleteMapping("/{id}")
     @PreAuthorize("hasRole('ADMIN')")
-    public ResponseEntity<Void> deleteBook(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteBook(@PathVariable Long id, Authentication auth) {
+        authService.ensureAdmin(auth.getName());
         bookService.deleteBook(id);
         return ResponseEntity.noContent().build();
     }
